+++ magik-checks/src/main/java/org/stevenlooman/sw/magik/MagikCheck.java
@@ -5,12 +5,10 @@ import com.sonar.sslr.api.Token;
 import org.sonar.check.Rule;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 public abstract class MagikCheck extends MagikVisitor {
   private List<MagikIssue> issues = new ArrayList<>();
@@ -38,10 +36,6 @@ public abstract class MagikCheck extends MagikVisitor {
    * @param column Column of issue.
    */
   public void addIssue(String message, int line, int column) {
-    if (isCheckDisabledAtLine(line)) {
-      return;
-    }
-
     issues.add(MagikIssue.lineColumnIssue(line, column, message, this));
   }
 
@@ -52,7 +46,7 @@ public abstract class MagikCheck extends MagikVisitor {
    */
   public void addIssue(String message, AstNode node) {
     Token token = node.getToken();
-    addIssue(message, token.getLine(), token.getColumn());
+    addIssue(message, token);
   }
 
   /**
@@ -70,34 +64,6 @@ public abstract class MagikCheck extends MagikVisitor {
    */
   public void addFileIssue(String message) {
     issues.add(MagikIssue.fileIssue(message, this));
-  }
-
-  private boolean isCheckDisabledAtLine(int lineNo) {
-    MagikVisitorContext context = getContext();
-    String fileContents = context.fileContent();
-    if (fileContents == null
-        || fileContents.isEmpty()) {
-      return false;
-    }
-
-    String[] lines = fileContents.split("\r\n|\n|\r");  // match BufferedReader.readLine()
-    String line = lines[lineNo - 1];
-    Pattern pattern = Pattern.compile(".*# ?mlint: ?disable=(.*)");
-    Matcher matcher = pattern.matcher(line);
-    if (!matcher.find()) {
-      return false;
-    }
-
-    String match = matcher.group(1);
-    List<String> disableds = Arrays.stream(match.split(","))
-        .map(x -> x.trim())
-        .collect(Collectors.toList());
-    String checkKey = this.getCheckKeyKebabCase();
-    if (disableds.contains(checkKey)) {
-      return true;
-    }
-
-    return false;
   }
 
   /**