+++ magik-squid/src/test/java/org/stevenlooman/sw/magik/analysis/scope/ScopeBuilderVisitorTest.java
@@ -8,6 +8,7 @@ import org.stevenlooman.sw.magik.MagikVisitorContext;
 import org.stevenlooman.sw.magik.parser.MagikParser;
 
 import java.nio.charset.Charset;
+import java.util.Arrays;
 
 public class ScopeBuilderVisitorTest {
 
@@ -27,7 +28,6 @@ public class ScopeBuilderVisitorTest {
         "\t_local a\n" +
         "_endmethod\n";
     MagikVisitorContext context = createContext(code);
-
     ScopeBuilderVisitor visitor = new ScopeBuilderVisitor();
     visitor.scanFile(context);
 
@@ -46,7 +46,6 @@ public class ScopeBuilderVisitorTest {
         "\t_local a, b\n" +
         "_endmethod\n";
     MagikVisitorContext context = createContext(code);
-
     ScopeBuilderVisitor visitor = new ScopeBuilderVisitor();
     visitor.scanFile(context);
 
@@ -69,7 +68,6 @@ public class ScopeBuilderVisitorTest {
         "\t_local (a, b) << (1, 2)\n" +
         "_endmethod\n";
     MagikVisitorContext context = createContext(code);
-
     ScopeBuilderVisitor visitor = new ScopeBuilderVisitor();
     visitor.scanFile(context);
 
@@ -92,7 +90,6 @@ public class ScopeBuilderVisitorTest {
         "\t_local a << b << _unset\n" +
         "_endmethod\n";
     MagikVisitorContext context = createContext(code);
-
     ScopeBuilderVisitor visitor = new ScopeBuilderVisitor();
     visitor.scanFile(context);
 
@@ -115,7 +112,6 @@ public class ScopeBuilderVisitorTest {
         "\t(a, b) << (1, 2)\n" +
         "_endmethod\n";
     MagikVisitorContext context = createContext(code);
-
     ScopeBuilderVisitor visitor = new ScopeBuilderVisitor();
     visitor.scanFile(context);
 
@@ -136,9 +132,8 @@ public class ScopeBuilderVisitorTest {
     String code =
         "_try _with a\n" +
         "_when error\n" +
-        "_endtry\n";
+        "_endtry";
     MagikVisitorContext context = createContext(code);
-
     ScopeBuilderVisitor visitor = new ScopeBuilderVisitor();
     visitor.scanFile(context);
 
@@ -154,12 +149,32 @@ public class ScopeBuilderVisitorTest {
   }
 
   @Test
+  public void testForLoop() {
+    String code =
+        "_for i, j _over a.fast_keys_and_elements()\n" +
+        "_loop\n" +
+        "_endloop";
+    MagikVisitorContext context = createContext(code);
+    ScopeBuilderVisitor visitor = new ScopeBuilderVisitor();
+    visitor.scanFile(context);
+
+    Scope globalScope = visitor.getGlobalScope();
+    Scope loopScope = globalScope.getSelfAndDescendantScopes().get(1);
+    ScopeEntry entryI = loopScope.getScopeEntry("i");
+    assertThat(entryI).isNotNull();
+    assertThat(entryI.getType() == ScopeEntry.Type.DEFINITION);
+
+    ScopeEntry entryJ = loopScope.getScopeEntry("j");
+    assertThat(entryJ).isNotNull();
+    assertThat(entryJ.getType() == ScopeEntry.Type.DEFINITION);
+  }
+
+  @Test
   public void testParameter() {
     String code =
         "_method object.m(a, _optional b)\n" +
-        "_endmethod\n";
+        "_endmethod";
     MagikVisitorContext context = createContext(code);
-
     ScopeBuilderVisitor visitor = new ScopeBuilderVisitor();
     visitor.scanFile(context);
 
@@ -179,9 +194,8 @@ public class ScopeBuilderVisitorTest {
   public void testParameterAssignment() {
     String code =
         "_method object.m << a\n" +
-        "_endmethod\n";
+        "_endmethod";
     MagikVisitorContext context = createContext(code);
-
     ScopeBuilderVisitor visitor = new ScopeBuilderVisitor();
     visitor.scanFile(context);
 
@@ -200,7 +214,6 @@ public class ScopeBuilderVisitorTest {
         "\t_return !current_grs! _is _unset\n" +
         "_endmethod";
     MagikVisitorContext context = createContext(code);
-
     ScopeBuilderVisitor visitor = new ScopeBuilderVisitor();
     visitor.scanFile(context);